+++ src/CoreLibrary.php
@@ -2,23 +2,13 @@
 
 namespace LaravelUi5\Core;
 
-use LaravelUi5\Core\Ui5\Contracts\Ui5LibraryInterface;
+use LaravelUi5\Core\Ui5\AbstractUi5Library;
 use LaravelUi5\Core\Enums\ArtifactType;
 use LaravelUi5\Core\Traits\HasAssetsTrait;
-use LaravelUi5\Core\Ui5\Contracts\Ui5ModuleInterface;
 
-class CoreLibrary implements Ui5LibraryInterface
+class CoreLibrary extends AbstractUi5Library
 {
     use HasAssetsTrait;
-
-    public function __construct(protected Ui5ModuleInterface $module)
-    {
-    }
-
-    public function getModule(): Ui5ModuleInterface
-    {
-        return $this->module;
-    }
 
     public function getType(): ArtifactType
     {