+++ tests/Fixture/.ui5-sources.php
@@ -1,0 +1,11 @@
+<?php
+
+return [
+    'modules' => [
+        'Hello' => [
+            'type' => 'app',
+            'path' => './',
+            'vendor' => 'Pragmatiqu IT GmbH',
+        ]
+    ]
+];